+++ jsonschema/_keywords.py
@@ -83,24 +83,6 @@ def items(validator, items, instance, schema):
             )
 
 
-def additionalItems(validator, aI, instance, schema):
-    if (
-        not validator.is_type(instance, "array")
-        or validator.is_type(schema.get("items", {}), "object")
-    ):
-        return
-
-    len_items = len(schema.get("items", []))
-    if validator.is_type(aI, "object"):
-        for index, item in enumerate(instance[len_items:], start=len_items):
-            yield from validator.descend(item, aI, path=index)
-    elif not aI and len(instance) > len(schema.get("items", [])):
-        error = "Additional items are not allowed (%s %s unexpected)"
-        yield ValidationError(
-            error % extras_msg(instance[len(schema.get("items", [])):]),
-        )
-
-
 def const(validator, const, instance, schema):
     if not equal(instance, const):
         yield ValidationError(f"{const!r} was expected")