+++ jsonschema/_legacy_keywords.py
@@ -101,6 +101,24 @@ def items_draft3_draft4(validator, items, instance, schema):
             )
 
 
+def additionalItems(validator, aI, instance, schema):
+    if (
+        not validator.is_type(instance, "array")
+        or validator.is_type(schema.get("items", {}), "object")
+    ):
+        return
+
+    len_items = len(schema.get("items", []))
+    if validator.is_type(aI, "object"):
+        for index, item in enumerate(instance[len_items:], start=len_items):
+            yield from validator.descend(item, aI, path=index)
+    elif not aI and len(instance) > len(schema.get("items", [])):
+        error = "Additional items are not allowed (%s %s unexpected)"
+        yield ValidationError(
+            error % _utils.extras_msg(instance[len(schema.get("items", [])):]),
+        )
+
+
 def items_draft6_draft7_draft201909(validator, items, instance, schema):
     if not validator.is_type(instance, "array"):
         return